+++ integration/setup/catalog.go
@@ -136,87 +136,3 @@ func installChart(ctx context.Context, config Config, releaseName, values, chart
 
 	return err
 }
-
-func installCertOperator(ctx context.Context, config Config) error {
-	certOperatorValues := `Installation:
-  V1:
-    Auth:
-      Vault:
-        Address: http://vault.default.svc.cluster.local:8200
-        Host: ""
-        CA:
-          TTL: 1440h
-        Certificate:
-          TTL: ""
-        Token:
-          TTL: ""
-        Version: ""
-    GiantSwarm:
-      CertOperator:
-        CRD:
-          LabelSelector: ""
-    Guest:
-      Calico:
-        CIDR: ""
-        Subnet: ""
-      Docker:
-        CIDR: ""
-      IPAM:
-        CIDRMask: ""
-        NetworkCIDR: ""
-        PrivateSubnetMask: ""
-        PublicSubnetMask: ""
-      Kubernetes:
-        API:
-          Auth:
-            Provider:
-              OIDC:
-                ClientID: ""
-                IssuerURL: ""
-                UsernameClaim: ""
-                GroupsClaim: ""
-          ClusterIPRange: ""
-          EndpointBase: k8s.%s
-        ClusterDomain: ""
-      SSH:
-        UserList: ""
-    Provider:
-      AWS:
-        Route53:
-          Enabled: false
-        S3AccessLogsExpiration: 0
-        TrustedAdvisor:
-          Enabled: false
-      Kind: ""
-    Secret:
-      CertOperator:
-        Service:
-          Vault:
-            Config:
-              Token: %s
-    Security:
-      RestrictAccess:
-        GuestAPI:
-          Public: false
-`
-	err := installLatestReleaseChartPackage(ctx, config, "cert-operator", fmt.Sprintf(certOperatorValues, env.CommonDomain(), env.VaultToken()))
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
-
-func installNodeOperator(ctx context.Context, config Config) error {
-	nodeOperatorValues := `Installation:
-  V1:
-    Registry:
-      Domain: quay.io
-`
-	err := installLatestReleaseChartPackage(ctx, config, "node-operator", nodeOperatorValues)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
